requirements.txt
Django
psycopg2
mysqlclient
django-dotenv
djangorestframework
django-debug-toolbar
Pillow
pyjwt
